# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API",
        "description": ""
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
...
    "paths": {
...
        "/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Cria um produto.",
                "description": "Cria um produto com os dados fornecidos.",
                "requestBody": {
                    "description": "Dados para criação de um produto.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "category"
                                ],
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Celular LG K10."
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Com o LG K10 4G na mão, não há nada que você não possa fazer Acesse as redes sociais, compartilhe fotos, jogue, assista vídeos e muito mais."
                                    },
                                    "price": {
                                        "type": "string",
                                        "example": 1240
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Eletronicos."
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Cria e retorna os dados do produto criado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Products"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou incompletos, ou já existentes.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "O requisitor não tem permissão para esse serviço.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Products"
                ],
                "description": "Retorna uma lista com todos os produtos.",
                "parameters": [
                    {
                        "name": "search_string",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "description": "Lista todos os produtos que contenham os caracteres fornecidos."
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "description": "Lista todos os produtos por categoria(as). É possivel fornecer mais de uma categoria, desde que sejam separadas por vírgula."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Produtos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Products"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou incompletos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "O requisitor não tem permissão para esse serviço.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "description": "Retorna todos os dados do produto com o respectivo ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna dados de um produto especificado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Products"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou incompletos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "O requisitor não tem permissão para esse serviço.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "description": "Atualiza os dados do produto com o respectivo ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Atualizado com sucesso.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou incompletos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "O requisitor não tem permissão para esse serviço.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "description": "Deleta o produto com o respectivo ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deletado com sucesso.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou incompletos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "O requisitor não tem permissão para esse serviço.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Products": {
                "schema": {
                    "required": [
                        "name",
                        "description",
                        "price",
                        "category"
                    ],
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string",
                            "example": "Celular LG K10."
                        },
                        "description": {
                            "type": "string",
                            "example": "Com o LG K10 4G na mão, não há nada que você não possa fazer Acesse as redes sociais, compartilhe fotos, jogue, assista vídeos e muito mais."
                        },
                        "price": {
                            "type": "string",
                            "example": 1240
                        },
                        "category": {
                            "type": "string",
                            "example": "Eletronicos."
                        }
                    }
                }
            },
            "Erro": {
                "description": "Mensagem de erro detalhada",
                "properties": {
                    "mensagem": {
                        "type": "object",
                        "description": "Conteúdo da mensagem de erro."
                    }
                }
            }
        }
    }
}